- allow higher resolution layers for Maxent and ALOC - speedup index creation
git-svn-id: http://alageospatialportal.googlecode.com/svn/trunk@345 73def43e-0e3b-11df-a5af-b38b39d06e06

+++ alaspatial/src/main/java/org/ala/spatial/analysis/index/SamplingIndex.java
@@ -182,7 +182,7 @@ public class SamplingIndex implements AnalysisIndexService {
                         TabulationSettings.environmental_data_path
                         + layer.name);
 
-                double[] values = grid.getValues2(points);
+                float[] values = grid.getValues2(points);
                
                 /* export values - RAF for writeDouble() */
                 RandomAccessFile raf = new RandomAccessFile(
